+++ lib/presentation/views/mode_selection_view.dart
@@ -1,5 +1,5 @@
+import 'package:cabo_counter/core/custom_theme.dart';
 import 'package:cabo_counter/l10n/app_localizations.dart';
-import 'package:cabo_counter/utility/custom_theme.dart';
 import 'package:flutter/cupertino.dart';
 
 class ModeSelectionMenu extends StatelessWidget {
