+++ lib/presentation/views/tab_view.dart
@@ -1,7 +1,7 @@
+import 'package:cabo_counter/core/custom_theme.dart';
 import 'package:cabo_counter/l10n/app_localizations.dart';
 import 'package:cabo_counter/presentation/views/information_view.dart';
 import 'package:cabo_counter/presentation/views/main_menu_view.dart';
-import 'package:cabo_counter/utility/custom_theme.dart';
 import 'package:flutter/cupertino.dart';
 
 class TabView extends StatefulWidget {
