+++ lib/presentation/widgets/stepper.dart
@@ -1,4 +1,4 @@
-import 'package:cabo_counter/utility/custom_theme.dart';
+import 'package:cabo_counter/core/custom_theme.dart';
 import 'package:flutter/cupertino.dart'; // Für iOS-Style
 
 class Stepper extends StatefulWidget {
